Read the unit types from a file
The unit type is a predefined value, which is declared to the unit_type.conf file, and used to define the unit of a Metric. The Accounting System reads that file for displaying to a client the available types.

The available types are cached for 24 hours. Afterward, the application reloads the file and displays the new unit types, if any.