Metric Types
The metric type is a predefined value, which is declared to the metric_type.conf file, and used to define the metric type of a Metric. The Accounting System reads that file for displaying to a client the available types.

The available types are cached for 24 hours. Afterward, the application reloads the file and displays the new metric types, if any.

ACC-17