Location in API responses
There is a proxy server acting on behalf of the Accounting System.
There are also many API responses containing the location of the created resources.
As a result, this location should encapsulate the proxy server URL.
To tackle that, we have to export a variable named SERVER_URL, which contains the proxy server URL.
The application reads the variable and incorporates it into every location.

ACC-28