Metric Definition Actions
The concept of role is introduced in this implementation. The Metric Definition roles are also defined.
In addition, every time a new entity is introduced into the Accounting System, it is linked to the unique feature we receive from the access token, the subject of the token.
This enables us to manage the degree of accessibility to the various entities.

ACC-32