Access Control
An access control entity is an entity that specifies which services
are granted access to particular entities. An individual entity can provide more
specific permissions that control access only to itself rather
than to the entire collection.

A mechanism has been created that allows the creation of Access Control in
all entities of the API. Only the Metric Definition mechanism
has been activated in this commit.

ACC-23