Hierarchical structure in the Accounting System
When creating and registering a Metric in an Installation, a hierarchical relationship is created among Project, Provider and Installation.
The root is the Project, the intermediate level is the Provider and finally the leaf is the Installation, which also contains the different Metrics.

For each different level of the above relationship you can request the corresponding Metrics.

ACC-52